refactor(user): corrected wrong filename
