+++ src/Contexts/shared/domain/interfaces/service.interface.ts
