+++ src/Contexts/user/application/get-user-by-id.service.ts
@@ -1,5 +1,5 @@
 import { UserRepository } from '../domain/repositories/user.repository'
-import { IService } from '../../shared/domain/interfaces/service.inteface'
+import { IService } from '../../shared/domain/interfaces/service.interface'
 import { UserResponseDTO } from '../domain/dtos/user-response.dto'
 
 export class GetUserByIdService implements IService<string, UserResponseDTO> {
